preparing for version change

+++ src/tet/data_process.py
@@ -32,18 +32,21 @@ def compress(zip_files, destination):
         shutil.make_archive(base_name=f"{destination}-zipped", format='zip')
         shutil.rmtree(path=destination)
 
-def createDir(destination):
-    try:
-        os.mkdir(destination)
-    except OSError as error:
-        print(error)
-        while True:
-            query = input("Directory exists, replace it? [y/n] ")
-            fl_1 = query[0].lower() 
-            if query == '' or not fl_1 in ['y','n']: 
-                print('Please answer with yes or no')
-            else:
-                shutil.rmtree(destination)
-                os.makedirs(destination)
-                break
-        if fl_1 == 'n': sys.exit(0)
+def createDir(destination, replace=True):
+    if replace:    
+        try:
+            os.mkdir(destination)
+        except OSError as error:
+            print(error)
+            while True:
+                query = input("Directory exists, replace it? [y/n] ")
+                fl_1 = query[0].lower() 
+                if query == '' or not fl_1 in ['y','n']: 
+                    print('Please answer with yes or no')
+                else:
+                    shutil.rmtree(destination)
+                    os.makedirs(destination)
+                    break
+            if fl_1 == 'n': sys.exit(0)
+    else:
+        os.makedirs(destination, exist_ok=True)
